Reduce warnings (#4381)
* Reduce warnings

* Use sbt-tpolecat

* Add missing compiler options

* Add doc scalac options

* Tune warnings

* split warn macro between 2.12 and 2.13

* Remove package-object-classes lint

* Update sbt-tpolecat version

* Remove skip-package

* Fix cast error

* Fix repl project scalac configuration

+++ scio-examples/src/main/scala/com/spotify/scio/examples/complete/game/LeaderBoard.scala
@@ -70,11 +70,11 @@ object LeaderBoard {
         .withZone(DateTimeZone.forTimeZone(TimeZone.getTimeZone("PST")))
     // Duration in minutes over which to calculate team scores, defaults to 1 hour
     val teamWindowDuration =
-      Duration.standardMinutes(args.int("teamWindowDuration", 60))
+      Duration.standardMinutes(args.long("teamWindowDuration", 60))
     // Data that comes in from our streaming pipeline after this duration isn't considered in our
     // processing. Measured in minutes, defaults to 2 hours
     val allowedLateness =
-      Duration.standardMinutes(args.int("allowedLateness", 120))
+      Duration.standardMinutes(args.long("allowedLateness", 120))
 
     // Read in streaming data from PubSub and parse each row as `GameActionInfo` events
     val gameEvents = sc
